requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]
python-multipart
pydantic-settings
python-dotenv
alembic
pytest
pytest-asyncio
httpx
